refactor(models): extract state CAS into shared state_cas feature
Secret and Receipt each carried a byte-identical copy of the atomic
compare-and-set primitive (STATE_CAS_SCRIPT + compare_and_set_state!) --
Secret in SecretStateManagement, Receipt in the misnamed DeprecatedFields.
Hoist the single copy into a shared Familia feature,
Onetime::Models::Features::StateCas, enabled on both models via
feature :state_cas, alongside the other cross-model features
(required_fields, with_migration_fields, passphrase_hashing).

Behaviour-preserving: the primitive is unchanged; only its home moves.
Both models now resolve compare_and_set_state! solely from the shared
feature (no local overrides remain), and the receipt TTL-resurrection
(#3625) and secret double-reveal race tryouts pass unchanged.
