+++ lib/onetime/models/receipt.rb
@@ -19,6 +19,7 @@ module Onetime
     feature :relationships
     feature :required_fields
     feature :housekeeping
+    feature :state_cas
     feature :deprecated_fields
 
     # Migration features - REMOVE after v1→v2 migration complete
