+++ lib/onetime/models/receipt/features/deprecated_fields.rb
@@ -271,55 +271,9 @@ module Onetime::Receipt::Features
         truncate.to_s == 'true'
       end
 
-      # Lua compare-and-set on the +state+ field, run atomically by Redis (its
-      # command loop is single-threaded). Sets state to ARGV[1] iff the current
-      # value equals one of ARGV[2..]. Returns 1 to the single caller that
-      # performs the flip, 0 to everyone else -- including when the key/field is
-      # gone (HGET yields a Lua false that matches nothing, so a TTL-evicted
-      # receipt is never resurrected) and when the state has already advanced
-      # (so a terminal receipt is never reverted). This mirrors the Secret's
-      # STATE_CAS_SCRIPT (SecretStateManagement) so both models share one
-      # concurrency idiom.
-      STATE_CAS_SCRIPT = <<~LUA
-        local current = redis.call('HGET', KEYS[1], 'state')
-        for i = 2, #ARGV do
-          if current == ARGV[i] then
-            redis.call('HSET', KEYS[1], 'state', ARGV[1])
-            return 1
-          end
-        end
-        return 0
-      LUA
-
-      private
-
-      # Atomically transition the persisted +state+ field from one of
-      # +from_states+ to +to_state+, returning whether THIS caller performed the
-      # flip. This is the single concurrency primitive behind every receipt state
-      # transition; each transition documents what its own guard protects.
-      #
-      # Because Redis runs the Lua script atomically, of any number of racing
-      # callers exactly one observes an allowed +from+ value and flips it. It
-      # fails closed: a missing key (TTL-evicted) or an already-advanced state
-      # simply loses, so a transition can neither resurrect an immortal, TTL-less
-      # receipt (#3625) nor revert a terminal one. On a win, the HSET lands on
-      # the live key, preserving its TTL.
-      #
-      # Operands are produced with #serialize_value so they match how +state+ is
-      # encoded at rest (Familia JSON-encodes scalar fields), and the eval uses
-      # the keyword +keys:+/+argv:+ form (as in claim_once! and the Secret CAS)
-      # so it does not depend on positional-argument compatibility across Redis
-      # client versions.
-      #
-      # @param to_state [Symbol, String] state to set on success.
-      # @param from_states [Array<Symbol, String>] states the flip may fire from.
-      # @return [Boolean] true iff this caller performed the transition.
-      def compare_and_set_state!(to_state, from_states)
-        argv = [serialize_value(to_state.to_s)]
-        from_states.each { |from_state| argv << serialize_value(from_state.to_s) }
-
-        dbclient.eval(STATE_CAS_SCRIPT, keys: [dbkey], argv: argv).to_i == 1
-      end
+      # The atomic +state+ compare-and-set every transition above claims through
+      # (+compare_and_set_state!+) is provided by the shared +state_cas+ feature;
+      # see Onetime::Models::Features::StateCas.
     end
   end
 end
